move tests from TagsCloudVisualization to new project TagsCloudVisualizationTests
