+++ cs/TagsCloudVisualizationTests/CircularCloudLayouterTests.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Drawing;
 using System.IO;
-using System.Linq;
 using NUnit.Framework;
 using FluentAssertions;
 using NUnit.Framework.Interfaces;
+using TagsCloudVisualization;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualizationTests
 {
     [TestFixture]
     public class CircularCloudLayouterTests
@@ -44,6 +44,7 @@ namespace TagsCloudVisualization
             Action putRectangle = () => CloudLayouter.PutNextRectangle(size);
 
             putRectangle.Should().Throw<ArgumentException>();
+            Assert.Throws<ArgumentException>(() => CloudLayouter.PutNextRectangle(size));
         }
 
         [Test]
