+++ cs/tdd.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples", "Samples\Samples.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{DE131E5A-19D8-4771-913C-DFD6F4F5B0B8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationTests", "TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj", "{D86F40AA-505D-4902-AF5F-40DEB269B926}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{DE131E5A-19D8-4771-913C-DFD6F4F5B0B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE131E5A-19D8-4771-913C-DFD6F4F5B0B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DE131E5A-19D8-4771-913C-DFD6F4F5B0B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D86F40AA-505D-4902-AF5F-40DEB269B926}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D86F40AA-505D-4902-AF5F-40DEB269B926}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D86F40AA-505D-4902-AF5F-40DEB269B926}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D86F40AA-505D-4902-AF5F-40DEB269B926}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
